pyproject.toml
[project]
name = "custom-mcp-server"
version = "0.1.0"
description = "MCP Memory Server - agent long-term memory backed by Lakebase + pgvector"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.12.5",
    "mcp[cli]>=1.14.0",
    "uvicorn>=0.34.2",
    "databricks-sdk>=0.60.0",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic>=2",
]

[project.scripts]
custom-mcp-server = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
